+++ src/graphnet/data/__init__.py
@@ -1,6 +1,7 @@
 """Modules for converting and ingesting data.
 
 `graphnet.data` enables converting domain-specific data to industry-standard,
-intermediate file formats  and reading this data.
+intermediate file formats and reading this data.
 """
-from .filters import I3Filter, I3FilterMask
+from .extractors.icecube.utilities.i3_filters import I3Filter, I3FilterMask
+from .dataconverter import DataConverter
