+++ src/graphnet/data/extractors/__init__.py
@@ -1,21 +1,2 @@
-"""Collection of I3Extractors, extracting pure-python data from I3Frames."""
-
-from .i3extractor import I3Extractor, I3ExtractorCollection
-from .i3featureextractor import (
-    I3FeatureExtractor,
-    I3FeatureExtractorIceCube86,
-    I3FeatureExtractorIceCubeDeepCore,
-    I3FeatureExtractorIceCubeUpgrade,
-    I3PulseNoiseTruthFlagIceCubeUpgrade,
-)
-from .i3truthextractor import I3TruthExtractor
-from .i3retroextractor import I3RetroExtractor
-from .i3splinempeextractor import I3SplineMPEICExtractor
-from .i3particleextractor import I3ParticleExtractor
-from .i3tumextractor import I3TUMExtractor
-from .i3hybridrecoextractor import I3GalacticPlaneHybridRecoExtractor
-from .i3genericextractor import I3GenericExtractor
-from .i3pisaextractor import I3PISAExtractor
-from .i3ntmuonlabelsextractor import I3NTMuonLabelExtractor
-from .i3quesoextractor import I3QUESOExtractor
+"""Module containing data-specific extractor modules."""
 from .extractor import Extractor
