+++ src/graphnet/data/extractors/icecube/i3retroextractor.py
@@ -2,8 +2,8 @@
 
 from typing import TYPE_CHECKING, Any, Dict
 
-from graphnet.data.extractors.i3extractor import I3Extractor
-from graphnet.data.extractors.utilities.frames import (
+from graphnet.data.extractors.icecube import I3Extractor
+from graphnet.data.extractors.icecube.utilities.frames import (
     frame_is_montecarlo,
     frame_is_noise,
 )
