+++ src/graphnet/data/extractors/icecube/utilities/collections.py
