--- src/graphnet/data/parquet/__init__.py
@@ -1,2 +1,0 @@
-"""Parquet-specific implementation of data classes."""
-from .parquet_dataconverter import ParquetDataConverter
