+++ src/graphnet/data/readers/__init__.py
@@ -1,0 +1,3 @@
+"""Modules for reading experiment-specific data and applying Extractors."""
+from .graphnet_file_reader import GraphNeTFileReader
+from .i3reader import I3Reader
