--- src/graphnet/data/sqlite/__init__.py
@@ -1,4 +1,0 @@
-"""SQLite-specific implementation of data classes."""
-from .sqlite_dataconverter import SQLiteDataConverter
-from .sqlite_utilities import create_table_and_save_to_sql
-from .sqlite_utilities import run_sql_code, save_to_sql
