+++ src/graphnet/data/utilities/__init__.py
@@ -1,1 +1,4 @@
 """Utilities for use across `graphnet.data`."""
+from .sqlite_utilities import create_table_and_save_to_sql
+from .sqlite_utilities import get_primary_keys
+from .sqlite_utilities import query_database
