+++ src/graphnet/data/utilities/parquet_to_sqlite.py
@@ -9,7 +9,9 @@ import numpy as np
 import pandas as pd
 from tqdm.auto import trange
 
-from graphnet.data.sqlite.sqlite_utilities import create_table_and_save_to_sql
+from graphnet.data.utilities.sqlite_utilities import (
+    create_table_and_save_to_sql,
+)
 from graphnet.utilities.logging import Logger
